Clean and sort robothub repository
