+++ src/basestation/LibusbUtilities.cpp
@@ -2,7 +2,7 @@
 // Created by emiel on 21-05-21.
 //
 
-#include "LibusbUtilities.h"
+#include "basestation/LibusbUtilities.h"
 
 #include <libusb-1.0/libusb.h>
